requirements.txt
aiobotocore==1.3.0
aiohttp==3.7.4.post0
aioitertools==0.7.1
alembic==1.6.2
altair==4.1.0
apache-airflow==2.1.0
apache-airflow-providers-apache-spark==1.0.3
apache-airflow-providers-ftp==1.1.0
apache-airflow-providers-google==3.0.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-postgres==1.0.2
apache-airflow-providers-sqlite==1.0.2
apispec==3.3.2
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==3.3.0
blinker==1.4
blis==0.7.4
boto3==1.17.84
botocore==1.20.49
bs4==0.0.1
cachetools==4.2.2
catalogue==2.0.4
cattrs==1.6.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
cloudpickle==1.6.0
coala-utils==0.5.1
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.13
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.5
dask==2021.5.1
decorator==5.0.9
decouple==0.0.7
defusedxml==0.7.1
demjson==2.2.4
dill==0.3.1.1
dnspython==1.16.0
docutils==0.17.1
email-validator==1.1.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
eventlet==0.31.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-AppBuilder==3.3.0
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fsspec==2021.5.0
gevent==21.1.2
gitdb==4.0.7
GitPython==3.1.17
google-ads==7.0.0
google-api-core==1.26.3
google-api-python-client==1.12.8
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-automl==2.3.0
google-cloud-bigquery==2.16.0
google-cloud-bigquery-datatransfer==3.1.1
google-cloud-bigquery-storage==2.4.0
google-cloud-bigtable==1.7.0
google-cloud-container==1.0.1
google-cloud-core==1.6.0
google-cloud-datacatalog==3.1.1
google-cloud-dataproc==2.3.1
google-cloud-dlp==1.0.0
google-cloud-kms==2.2.0
google-cloud-language==1.3.0
google-cloud-logging==2.3.1
google-cloud-memcache==0.3.0
google-cloud-monitoring==2.2.1
google-cloud-os-login==2.1.0
google-cloud-pubsub==2.4.2
google-cloud-redis==2.1.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.38.0
google-cloud-tasks==2.2.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==0.3.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
graphviz==0.16
greenlet==1.1.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.1
grpcio-gcp==0.2.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.3
httplib2==0.17.4
httpx==0.18.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
jtutils==0.0.8
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
leven==1.0.4
libcst==0.3.18
locket==0.2.1
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
MechanicalSoup==1.1.0
mistune==0.8.4
multidict==5.1.0
murmurhash==1.0.5
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.0
numpy==1.20.3
oauthlib==2.1.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
packaging==20.9
pandas==1.2.4
pandas-gbq==0.14.1
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.2
partd==1.2.0
pathy==0.5.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
plyvel==1.3.0
preshed==3.0.5
prison==0.1.3
prometheus-client==0.11.0
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.16.0
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py4j==0.10.9
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.4
pydata-google-auth==1.2.0
pydeck==0.6.2
pyee==8.1.0
Pygments==2.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
pyrsistent==0.17.3
PySocks==1.7.1
pyspark==3.1.1
python-csv==0.0.13
python-daemon==2.3.0
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.1.0
rfc3986==1.5.0
rich==9.2.0
rsa==4.7.2
s3fs==2021.5.0
s3transfer==0.4.2
selenium==3.141.0
Send2Trash==1.5.0
setproctitle==1.2.2
six==1.16.0
smart-open==3.0.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.2
srsly==2.4.1
streamlit==0.82.0
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
twitterscraper==1.6.1
typer==0.3.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
validators==0.18.2
w3lib==1.22.0
wasabi==0.8.2
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.6.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
